refactor: 안쓰는 코드 정리
- signup success Link 새창으로 연결
- 모달 이름 및 폴더 이름 변경
- 필요없는 주석 삭제
