+++ app/(shared)/atoms/modals.atom.ts
@@ -1,4 +1,3 @@
-// modals.atom.ts
 import {
   atom,
   useRecoilState,
