+++ app/(shared)/components/Dialog/Noti/DialogBody.tsx
