+++ app/(shared)/components/Dialog/TimerImageDelete/Dialog.tsx
