+++ app/(shared)/components/Dialog/TimerImageDelete/DialogBody.tsx
