+++ app/(shared)/components/Dialog/dialog.sass
@@ -2,8 +2,6 @@
 @import '@/(shared)/styles/mixins'
 @import '@/(shared)/styles/button'
 
-
-// 변수 정의
 .theme-info-modal
   position: fixed
   top: 50%
@@ -20,7 +18,7 @@
     align-items: center
     margin-bottom: 12px
     &.noti
-      justify-content: center  // 중앙 정렬  
+      justify-content: center
     h2
       @include title20SB
       cursor: default
@@ -216,7 +214,7 @@
     padding: 24px 24px 12px 32px
     height: 68px
     &.noti
-      justify-content: center  // 중앙 정렬  
+      justify-content: center
     h2
       @include title20SB
       cursor: default
@@ -311,7 +309,7 @@
     padding: 24px 24px 12px 32px
     height: 68px
     &.noti
-      justify-content: center  // 중앙 정렬  
+      justify-content: center
     h2
       @include title20SB
       cursor: default
