+++ app/(shared)/components/ThemeTextField/textField.modules.sass
@@ -59,11 +59,6 @@
     transform: translateY(10px)
   to
     transform: translateY(0)
-// @keyframes fadeOut
-//   from 
-//     transform: translateY(10px)
-//   to
-//     transform: translateY(0)
 
 .theme-textfield-text
   margin: 4px 0 0 16px
