+++ app/(shared)/components/Toast/toast.sass
@@ -25,7 +25,6 @@
 .toast-body
   color: $color-black
 
-// 애니메이션 효과
 @keyframes fadeIn
   from
     opacity: 0
@@ -36,7 +35,6 @@
 
 .toast-message
   animation: fadeIn 0.3s ease-out forwards
-// 반응형 디자인
 @media (max-width: 480px)
   .toast-container
     position: fixed
